% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiBED_deconvolution.R
\name{HiBED_deconvolution}
\alias{HiBED_deconvolution}
\title{HiBED_deconvolution}
\usage{
HiBED_deconvolution(Beta, h = 2)
}
\arguments{
\item{Beta}{Methylation beta matrix from brain samples.}

\item{h}{Specify the layer of deconvolution in the hierarchical model. Default is 2.}
}
\value{
A matrix with predicted cell proportions in brain tissues.
}
\description{
The function estimates proportions up to 7 cell types in brain tissues.
}
